+++ contracts/__init__.py
@@ -6,10 +6,12 @@ from .i_service import IService
 from .i_file_system import IFileSystem
 from .i_event_bus import IEventBus
 from .i_capability_registry import ICapabilityRegistry
+from .igraph_builder import IGraphBuilder
 
 __all__ = [
     "IService",
     "IFileSystem",
     "IEventBus",
     "ICapabilityRegistry",
+    "IGraphBuilder",
 ]